requirements.txt
fastapi
praw
python-dotenv
pandas
